SHM5 SHM6 - Coordinates class implemented inside Island
