+++ shm/island.hpp
@@ -1,8 +1,18 @@
 #pragma once
-#include "coordinates.hpp"
 
 class Island {
 public:
+    class Coordinates {
+    public:
+        Coordinates() = default;
+        Coordinates(int positionX, int positionY);
+        bool operator==(const Coordinates& rhs) const;
+
+    private:
+        int positionX_{0};
+        int positionY_{0};
+    };
+
     Island(int position_X, int position_Y);
     Coordinates getPosition() const { return position_; }
